Move first functions to new package
