+++ perun/collect/trace/ebpf/engine.py
@@ -11,6 +11,7 @@ import perun.collect.trace.ebpf.program as program
 import perun.logic.temp as temp
 import perun.utils.metrics as metrics
 from perun.collect.trace.watchdog import WATCH_DOG
+from perun.utils.external import environment
 
 try:
     import bcc
@@ -84,7 +85,7 @@ class BpfEngine(engine.CollectEngine):
 
         WATCH_DOG.info("Starting up the eBPF collection process.")
         # Run the new ebpf process with sudo privileges
-        current_interpreter = utils.get_current_interpreter("3.6+")
+        current_interpreter = environment.get_current_interpreter("3.6+")
         with utils.nonblocking_subprocess(
             f"sudo {current_interpreter} {_get_ebpf_file()} {self.runtime_conf}",
             {},
