+++ perun/utils/external/meson.build
@@ -1,0 +1,11 @@
+perun_external_dir = perun_utils_dir / 'external'
+
+perun_external_files = files(
+    '__init__.py',
+    'environment.py',
+)
+
+py3.install_sources(
+    perun_external_files,
+    subdir: perun_external_dir
+)
